Separate providers into files
* Changes can be tracked easier on a per provider basis.
* fix trailing comma with intercom
* update makefile to use google closure compiler to build
* built files go into dist folder
* update tests to use built version
