+++ Makefile
@@ -1,5 +1,13 @@
-min:
-	uglifyjs -o analytics.min.js analytics.js
+dist: dist/analytics.min.js
+
+dist/analytics.js: analytics.js
+	reunion --ns analytics $< > $@
+
+# minification courtesy of googlz
+dist/analytics.min.js: dist/analytics.js
+	curl --data-urlencode "js_code@$<" \
+		-d "output_info=compiled_code&compilation_level=SIMPLE_OPTIMIZATIONS" \
+		http://closure-compiler.appspot.com/compile > $@
 
 test:
 	open test/analyticsjs.html
