+++ providers/index.js
@@ -1,0 +1,12 @@
+module.exports = {
+    'Google Analytics': require('./google'),
+    'KISSmetrics': require('./kiss'),
+    'Mixpanel': require('./mixpanel'),
+    'Intercom': require('./intercom'),
+    'HubSpot': require('./hubspot'),
+    'Chartbeat': require('./chartbeat'),
+    'Olark': require('./olark'),
+    'CrazyEgg': require('./crazyegg'),
+    'Customer.io': require('./customer')
+};
+
