+++ test/providers.html
@@ -16,7 +16,7 @@
         window.expect = window.chai.expect;
     </script>
 
-    <script src="../analytics.js"></script>
+    <script src="../dist/analytics.js"></script>
     <script src="providers.test.js"></script>
 </head>
 <body>
